Separate NotificationBadgeWithPopover from NotificationBadge
