+++ src/components/assessment/index.tsx
@@ -25,7 +25,7 @@ import { NavLink } from 'react-router-dom';
 import { sortBy } from 'lodash';
 import defaultCoverImage from '../../assets/default_cover_image.jpg';
 import AssessmentWorkspaceContainer from '../../containers/assessment/AssessmentWorkspaceContainer';
-import NotificationBadge from '../../containers/notification/NotificationBadge';
+import NotificationBadgeWithPopover from '../../containers/notification/NotificationBadgeWithPopover';
 import { beforeNow, getPrettyDate } from '../../utils/dateHelpers';
 import { assessmentCategoryLink, stringParamToInt } from '../../utils/paramParseHelpers';
 import {
@@ -334,7 +334,11 @@ const makeOverviewCard = (
   <div key={index}>
     <Card className="row listing" elevation={Elevation.ONE}>
       <div className="col-xs-3 listing-picture">
-        <NotificationBadge className="badge" notifications={notifications} large={true} />
+        <NotificationBadgeWithPopover
+          className="badge"
+          large={true}
+          notifications={notifications}
+        />
         <img
           className={`cover-image-${overview.status}`}
           src={overview.coverImage ? overview.coverImage : defaultCoverImage}
