+++ src/components/notification/__tests__/NotificationBadge.tsx
@@ -46,38 +46,21 @@ const avengerNotifications: Notification[] = [
   }
 ];
 
-const handleAcknowledgeNotifications = (ids: number[]) => null;
-
 describe('Badge', () => {
   test('renders properly for students', () => {
-    const tree = mount(
-      <NotificationBadge
-        notifications={studentNotifications}
-        handleAcknowledgeNotifications={handleAcknowledgeNotifications}
-      />
-    );
+    const tree = mount(<NotificationBadge number={studentNotifications.length} />);
 
     expect(tree.debug()).toMatchSnapshot();
   });
 
   test('renders properly for avengers', () => {
-    const tree = mount(
-      <NotificationBadge
-        notifications={avengerNotifications}
-        handleAcknowledgeNotifications={handleAcknowledgeNotifications}
-      />
-    );
+    const tree = mount(<NotificationBadge number={avengerNotifications.length} />);
 
     expect(tree.debug()).toMatchSnapshot();
   });
 
   test('does not render with no notifications', () => {
-    const tree = mount(
-      <NotificationBadge
-        notifications={[]}
-        handleAcknowledgeNotifications={handleAcknowledgeNotifications}
-      />
-    );
+    const tree = mount(<NotificationBadge number={0} />);
 
     expect(tree.debug()).toMatchSnapshot();
   });
